Add explicit declaration of static constexpr arrays in reconstruct helper class lists to avoid link failures.  These are declared in new file instantiate.cu
